Remove hard dependency on postgresql
The only part where we using postgresql specific fields was on the
link relation types. It could be removed by add a proper LinkRelation
model table, and then having the 'relations' as a property that checks
the associated types.

PostgreSQL is still required however, because sqlite is not supported
by django-ulid.